# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "store-api overview",
    "version": "v1"
  },
  "paths": {
    "/": {
      "get": {
        "responses": {
          "301": {
            "description": "Nothing to see here"
          }
        }
      }
    },
    "/api-docs": {
      "get": {
        "summary": "This site",
        "responses": {
          "200": {
            "description": "Swagger-Ui for this API"
          }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "List all products",
        "tags": [
          "product"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "from",
            "type": "integer",
            "in": "query",
            "required": false
          },
          {
            "name": "to",
            "type": "integer",
            "in": "query",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "List of all products"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get details of a product",
        "tags": [
          "product"
        ],
        "parameters": [
          {
            "name": "id",
            "type": "integer",
            "in": "path",
            "required": true
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Product object"
          },
          "404": {
            "description": "No Product object for this id"
          }
        }
      }
    },
    "/products/search/{query}": {
      "get": {
        "summary": "Get all products which contains the query in their description",
        "tags": [
          "product"
        ],
        "description": "Checks if the description of the product contains the query -> ⚠️ object.description.contains(query)⚠️",
        "parameters": [
          {
            "name": "query",
            "type": "string",
            "in": "path",
            "required": true
          },
          {
            "name": "from",
            "type": "integer",
            "in": "query",
            "required": false
          },
          {
            "name": "to",
            "type": "integer",
            "in": "query",
            "required": false
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Array of Products containing the query"
          }
        }
      }
    },
    "/price/{id}": {
      "get": {
        "summary": "Get price of a product",
        "description": "Use the 'duration' query param to get the price for a specified duration. ",
        "tags": [
          "pricing"
        ],
        "parameters": [
          {
            "name": "id",
            "type": "integer",
            "in": "path",
            "required": true
          },
          {
            "name": "duration",
            "type": "integer",
            "in": "query",
            "required": false,
            "description": "Duration in months"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Price object"
          },
          "404": {
            "description": "No Price object for this id"
          }
        }
      }
    },
    "/add-to-cart": {
      "post": {
        "tags": [
          "not ready"
        ],
        "responses": {
          "501": {
            "description": "Coming Soon"
          }
        }
      }
    },
    "/cart/{id}": {
      "get": {
        "tags": [
          "not ready"
        ],
        "responses": {
          "501": {
            "description": "Coming Soon"
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": [
          "not ready"
        ],
        "responses": {
          "501": {
            "description": "Coming Soon"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "not ready"
        ],
        "responses": {
          "501": {
            "description": "Coming Soon"
          }
        }
      }
    }
  },
  "consumes": [
    "application/json"
  ]
}